% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cytof_postProcess.R
\name{add_col_to_fcs}
\alias{add_col_to_fcs}
\title{Add new columns to the fcs expression data}
\usage{
add_col_to_fcs(data, rawFCSdir, analyzedFCSdir, transformed_col = c("tsne_1",
  "tsne_2"), cluster_col = c("cluster"))
}
\arguments{
\item{data}{The new data matrix to be added in.}

\item{rawFCSdir}{The directory containing the original fcs files.}

\item{analyzedFCSdir}{The directory to store the new fcs files.}

\item{transformed_col}{the column name of the dimension transformend data in \code{data}.}

\item{cluster_col}{the column name of the cluster data in \code{data}.}
}
\value{
new fcs files stored under \code{analyzedFCSdir}
}
\description{
Store the new dimension transformed data and cluster data into the exprs
matrix in new fcs files under \code{analyzedFCSdir}
}

