% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_preProcess.R
\name{applyComp}
\alias{applyComp}
\title{apply compensation on the FCS expression data}
\usage{
applyComp(fcs, compMatrix)
}
\arguments{
\item{fcs}{FCS file.}

\item{keyword}{Keywords.}
}
\value{
compensated expression value
}
\description{
apply compensation on the FCS expression data
}

