% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cytof_preProcess.R
\name{auto_lgcl}
\alias{auto_lgcl}
\title{a modified version of "estimateLogicle" from flowCore}
\usage{
auto_lgcl(x, channels, m = 4.5, q = 0.05)
}
\arguments{
\item{x}{Data.}

\item{channels}{Channel names.}

\item{m}{Para m.}

\item{q}{Para q.}
}
\description{
a modified version of "estimateLogicle" from flowCore
}

