% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_preProcess.R
\name{autoLgcl}
\alias{autoLgcl}
\title{a modified version of "estimateLogicle" from flowCore}
\usage{
autoLgcl(x, channels, m = 4.5, q = 0.05)
}
\arguments{
\item{x}{Data.}

\item{channels}{Channel names.}

\item{m}{Para m.}

\item{q}{Para q.}
}
\value{
a list of transformations
}
\description{
a modified version of "estimateLogicle" from flowCore
}

