% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cytof_postProcess.R
\name{clust_percentage_heatmap}
\alias{clust_percentage_heatmap}
\title{Heatmap plot of cluster percentage results}
\usage{
clust_percentage_heatmap(clust_cellCount, baseName = "Cluster percentage",
  scaleMethod = "none")
}
\arguments{
\item{clust_cellCount}{cluster count data from results of \code{clust_state}}

\item{baseName}{The name as a prefix in the title of the heatmap}

\item{scaleMethod}{character indicating if the values should be centered and scaled in either the row direction or the column direction, or none. The default is "none".}
}
\value{
a heatmap object from \code{gplots}
}
\description{
Heatmap plot of cluster percentage results
}
\examples{
m1 <- c(rnorm(300, 10, 2), rnorm(400, 4, 2), rnorm(300, 7))
m2 <- c(rnorm(300, 4), rnorm(400, 16), rnorm(300, 10, 3))
m3 <- c(rnorm(300, 16), rnorm(400, 40, 3), rnorm(300, 10))
m4 <- c(rnorm(300, 7, 3), rnorm(400, 30, 2), rnorm(300, 10))
m5 <- c(rnorm(300, 27), rnorm(400, 40, 1),rnorm(300, 10))
c <- c(rep(1,300), rep(2,400), rep(3,300))
rnames <- paste(paste('sample_', c('A','B','C','D'), sep = ''),
rep(1:250,each = 4), sep='_')
exprs_cluster <- data.frame(cluster = c, m1 = m1, m2 = m2, m3 = m3, m4 = m4, m5 = m5)
row.names(exprs_cluster) <- sample(rnames, 1000)
clust_statData <- clust_state(exprs_cluster)
clust_percentage_heatmap(clust_statData[[2]])
}

