% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cytof_postProcess.R
\name{clust_state}
\alias{clust_state}
\title{Statistical analysis of the cluster results}
\usage{
clust_state(exprs_cluster, stat = "mean")
}
\arguments{
\item{exprs_cluster}{the expression matrix combined with the cluster results}

\item{stat}{the method used for statistical analysis, like mean, median...}
}
\value{
a list contains a matrix of \code{clust_mean} and a matrix of \code{clust_cellCount}
}
\description{
Calculate the mean value of each markers in each cluster, If there are multiple samples, the percentage
of cells in each cluster in each sample will be calculated
}
\examples{
m1 <- c(rnorm(300, 10, 2), rnorm(400, 4, 2), rnorm(300, 7))
m2 <- c(rnorm(300, 4), rnorm(400, 16), rnorm(300, 10, 3))
m3 <- c(rnorm(300, 16), rnorm(400, 40, 3), rnorm(300, 10))
m4 <- c(rnorm(300, 7, 3), rnorm(400, 30, 2), rnorm(300, 10))
m5 <- c(rnorm(300, 27), rnorm(400, 40, 1),rnorm(300, 10))
c <- c(rep(1,300), rep(2,400), rep(3,300))
rnames <- paste(paste('sample_', c('A','B','C','D'), sep = ''),
rep(1:250,each = 4), sep='_')
exprs_cluster <- data.frame(cluster = c, m1 = m1, m2 = m2, m3 = m3, m4 = m4, m5 = m5)
row.names(exprs_cluster) <- sample(rnames, 1000)
clust_statData <- clust_state(exprs_cluster)
}

