% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cytof_postProcess.R
\name{cluster_plot}
\alias{cluster_plot}
\title{Scatter plot of the cluster results}
\usage{
cluster_plot(clusterData, title = "cluster", point_size = NULL)
}
\arguments{
\item{clusterData}{The matrix of cluster results, with rownames of their sample name and cell id.}

\item{title}{the title name of the plot.}

\item{point_size}{the size of the dot.}
}
\value{
the scatter dot plot
}
\description{
Dot plot visualization of the cluster results, with color indicating different clusters,
and shape of different samples.
}
\examples{
x <- c(rnorm(100, mean = 1), rnorm(100, mean = 3), rnorm(100, mean = 9))
y <- c(rnorm(100, mean = 2), rnorm(100, mean = 8), rnorm(100, mean = 5))
c <- c(rep(1,100), rep(2,100), rep(3,100))
rnames <- paste(paste('sample_', c('A','B','C'), sep = ''), rep(1:100,each = 3), sep='_')
clusterData <- data.frame(dim1 = x, dim2 = y, cluster = c)
rownames(clusterData) <- rnames
cluster_plot(clusterData)
}

