% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterX.R
\name{ClusterX}
\alias{ClusterX}
\title{Fast clustering by automaticly search and find of density peaks}
\usage{
ClusterX(data, dimReduction = NULL, outDim = 2, dc, gaussian = TRUE,
  alpha = 0.001, detectHalos = FALSE, parallel = FALSE, nCore = 4)
}
\arguments{
\item{data}{A data matrix for clustering.}

\item{dimReduction}{Dimenionality reduciton method.}

\item{outDim}{Number of dimensions will be used.}

\item{dc}{Distance cutoff value.}

\item{gaussian}{If apply gaussian to esitmate the density.}

\item{alpha}{Signance level for peak detection.}

\item{detectHalos}{If detect the halos.}

\item{parallel}{If run the algorithm in parallel.}

\item{nCore}{Number of cores umployed for parallel compution.}
}
\value{
a object of \code{ClusterX} class
}
\description{
This package implement the clustering algorithm described by Alex Rodriguez
and Alessandro Laio (2014) with improvements of automatic peak detection and 
parallel implementation
}
\examples{
iris_unique <- unique(iris) # Remove duplicates
data <- as.matrix(iris_unique[,1:4])
ClusterXRes <- ClusterX(data)
}
\author{
Chen Hao
}

