% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_postProcess.R
\name{cytof_addToFCS}
\alias{cytof_addToFCS}
\title{Add data to the original FCS files}
\usage{
cytof_addToFCS(data, rawFCSdir, analyzedFCSdir, transformed_cols = c("tsne_1",
  "tsne_2"), cluster_cols = c("cluster"), inLgclTrans = TRUE)
}
\arguments{
\item{data}{The new data matrix to be added in.}

\item{rawFCSdir}{The directory containing the original fcs files.}

\item{analyzedFCSdir}{The directory to store the new fcs files.}

\item{transformed_cols}{the column name of the dimension transformend data in \code{data}.}

\item{cluster_cols}{the column name of the cluster data in \code{data}.}

\item{inLgclTrans}{Boolean value decides if apply the inverse lgcl transformation to the data before saving}
}
\value{
new fcs files stored under \code{analyzedFCSdir}
}
\description{
Store the new dimension transformed data and cluster data into the exprs 
matrix in new fcs files under \code{analyzedFCSdir}
}

