% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_cluster.R
\name{cytof_cluster}
\alias{cytof_cluster}
\title{Subset detection by clustering}
\usage{
cytof_cluster(ydata = NULL, xdata = NULL, method = c("Rphenograph",
  "ClusterX", "DensVM", "FlowSOM", "NULL"), FlowSOM_k = 40)
}
\arguments{
\item{ydata}{A matrix of the dimension reduced data.}

\item{xdata}{A matrix of the expression data.}

\item{method}{Cluster method including \code{DensVM}, \code{densityClustX}, \code{Rphenograph} and \code{FlowSOM}.}

\item{FlowSOM_k}{Number of clusters for meta clustering in FlowSOM.}
}
\value{
a vector of the clusters assigned for each row of the ydata
}
\description{
Apply clustering algorithms to detect cell subsets. \code{DensVM} and \code{ClusterX} 
clustering is based on the transformed ydata and uses xdata to train the model. 
\code{Rphenograph} directly works on high dimensional xdata. \code{FlowSOM} is 
integrated from FlowSOM pacakge (https://bioconductor.org/packages/release/bioc/html/FlowSOM.html).
}
\examples{
d<-system.file('extdata', package='cytofkit')
fcsFile <- list.files(d, pattern='.fcs$', full=TRUE)
parameters <- list.files(d, pattern='.txt$', full=TRUE)
markers <- as.character(read.table(parameters, header = FALSE)[, 1])
xdata <- cytof_exprsMerge(fcsFile, mergeMethod = 'fixed', fixedNum = 100)
ydata <- cytof_dimReduction(xdata, markers = markers, method = "tsne")
clusters <- cytof_cluster(ydata, xdata, method = "ClusterX")
}
