% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_postProcess.R
\name{cytof_clusterMtrx}
\alias{cytof_clusterMtrx}
\title{Expression values by cluster}
\usage{
cytof_clusterMtrx(analysis_results, clusterMethod = NULL, cluster)
}
\arguments{
\item{analysis_results}{analysis_results output of cytofkit to extract expression values from.}

\item{clusterMethod}{The cluster method of interest. If none is selected, the first one will be used.}

\item{cluster}{The cluster of interest.}
}
\value{
A data table with marker expression values of cells in a given cluster
}
\description{
Generate a matrix of expression values for cells in a cluster
}
\examples{
 
dir <- system.file('extdata',package='cytofkit')
file <- list.files(dir ,pattern='.RData$', full=TRUE)
load(file)
cluster2_table <- cytof_clusterMtrx(analysis_results, "Rphenograph", 2)
}
