% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cytof_postProcess.R
\name{cytof_clusterPlot}
\alias{cytof_clusterPlot}
\title{Scatter plot of the cluster results}
\usage{
cytof_clusterPlot(data, xlab, ylab, cluster, sample, title = "cluster",
  type = 1, point_size = NULL)
}
\arguments{
\item{data}{The data frame of cluster results}

\item{xlab}{The name for the x axis in input \code{data}}

\item{ylab}{The name for the y axis in input \code{data}}

\item{sample}{the name of the sample column in input \code{data}}

\item{title}{the title name of the plot.}

\item{type}{plot type, 1 indicates combined plot, 2 indicated grid facet plot by samples}

\item{point_size}{the size of the dot.}

\item{clueter}{the name of cluter column in input \code{data}}
}
\value{
the ggplot object of the scatter dot plot
}
\description{
Dot plot visualization of the cluster results, with color indicating different clusters,
and shape of different samples.
}
\examples{
x <- c(rnorm(100, mean = 1), rnorm(100, mean = 3), rnorm(100, mean = 9))
y <- c(rnorm(100, mean = 2), rnorm(100, mean = 8), rnorm(100, mean = 5))
c <- c(rep(1,100), rep(2,100), rep(3,100))
rnames <- paste(paste('sample_', c('A','B','C'), sep = ''), rep(1:100,each = 3), sep='_')
data <- data.frame(dim1 = x, dim2 = y, cluster = c)
rownames(data) <- rnames
cytof_clusterPlot(data)
}

