% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_postProcess.R
\name{cytof_clusterStat}
\alias{cytof_clusterStat}
\title{Statistics of the cluster results}
\usage{
cytof_clusterStat(data, markers, cluster = "cluster", sample,
  statMethod = c("mean", "median", "percentage", "NULL"))
}
\arguments{
\item{data}{Input data frame.}

\item{markers}{The names of markers used for calcualtion.}

\item{cluster}{The column name contatining cluster labels.}

\item{sample}{The samples used for calculation.}

\item{statMethod}{Statistics containing mean, median or percentage.}
}
\value{
A matrix of the statistics results
}
\description{
Calculate the mean or median expression level of each marker for each cluster, or percentage
of cell numbers of each cluster for each sample.
}
\examples{
m1 <- c(rnorm(300, 10, 2), rnorm(400, 4, 2), rnorm(300, 7))
m2 <- c(rnorm(300, 4), rnorm(400, 16), rnorm(300, 10, 3))
m3 <- c(rnorm(300, 16), rnorm(400, 40, 3), rnorm(300, 10))
m4 <- c(rnorm(300, 7, 3), rnorm(400, 30, 2), rnorm(300, 10))
m5 <- c(rnorm(300, 27), rnorm(400, 40, 1),rnorm(300, 10))
c <- c(rep(1,300), rep(2,400), rep(3,300))
rnames <- paste(paste('sample_', c('A','B','C','D'), sep = ''), 
rep(1:250,each = 4), sep='_') 
exprs_cluster <- data.frame(cluster = c, m1 = m1, m2 = m2, m3 = m3, m4 = m4, m5 = m5)
row.names(exprs_cluster) <- rnames
cytof_clusterStat(data = exprs_cluster, cluster = "cluster", statMethod = "mean")
}
