% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_postProcess.R
\name{cytof_colorPlot}
\alias{cytof_colorPlot}
\title{Plot the data with color-coded marker values}
\usage{
cytof_colorPlot(data, xlab, ylab, zlab, colorPalette = c("bluered",
  "spectral1", "spectral2", "heat"), pointSize = 1, removeOutlier = TRUE)
}
\arguments{
\item{data}{A dataframe containing the xlab, ylab and zlab.}

\item{xlab}{The column name of data for x lab.}

\item{ylab}{The column name of data for y lab.}

\item{zlab}{The column name of data for z lab.}

\item{colorPalette}{Color Palette.}

\item{pointSize}{Size of the point.}

\item{removeOutlier}{If \verb{TRUE}, remove the outliers.}
}
\value{
A ggplot object.
}
\description{
Plot the data with color-coded marker values
}
\examples{
x <- c(rnorm(100, mean = 1), rnorm(100, mean = 3), rnorm(100, mean = 9))
y <- c(rnorm(100, mean = 2), rnorm(100, mean = 8), rnorm(100, mean = 5))
c <- rnorm(300, 10, 5)
data <- data.frame(dim1 = x, dim2 = y, marker = c)
cytof_colorPlot(data = data, xlab = "dim1", ylab = "dim2", zlab = "marker")
}
