% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cytof_dimensionReduction.R
\name{cytof_dimReduction}
\alias{cytof_dimReduction}
\title{Dimension reduction of CyTof expression data}
\usage{
cytof_dimReduction(data, method = "tsne", distMethod = "euclidean",
  isomap_k = 5, isomap_ndim = NULL, isomapFragmentOK = TRUE)
}
\arguments{
\item{data}{An expression data matrix}

\item{method}{Method chosed for dimensition reduction, must be one of \code{isomap}, \code{pca} or \code{tsne}}

\item{distMethod}{Method for distance calcualtion}

\item{isomap_k}{Number of shortest dissimilarities retained for a point, parameter for \code{isomap} method}

\item{isomap_ndim}{Number of axes in metric scaling, parameter for \code{isomap} method}

\item{isomapFragmentOK}{What to do if dissimilarity matrix is fragmented, parameter for \code{isomap} method}
}
\value{
a matrix of the dimension reducted data, with colnames and rownames(if have, same as the input)
}
\description{
Apply dimension reduction on the CyTof expression data,
with method \code{isomap}, \code{pca}, or \code{tsne}.
}
\examples{
data(iris)
in_data <- iris[, 1:4]
out_data <- cytof_dimReduction(in_data)
}
\author{
Chen Jinmiao
}

