% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_preProcess.R
\name{cytof_exprsMerge}
\alias{cytof_exprsMerge}
\title{Merge the expression matrix from multiple FCS files with preprocessing}
\usage{
cytof_exprsMerge(fcsFiles, comp = FALSE, transformMethod = c("autoLgcl",
  "cytofAsinh", "logicle", "arcsinh", "none"), scaleTo = NULL,
  mergeMethod = c("ceil", "all", "fixed", "min"), fixedNum = 10000,
  sampleSeed = 123, ...)
}
\arguments{
\item{fcsFiles}{A vector of FCS file names.}

\item{comp}{If \verb{TRUE}, does compensation  by compensation matrix contained in FCS. Agrument also accepts a compensation matrix to be applied. Otherwise \verb{FALSE}.}

\item{transformMethod}{Data Transformation method, including \code{autoLgcl}, \code{cytofAsinh}, \code{logicle} and \code{arcsinh}, or \code{none} to avoid transformation.}

\item{scaleTo}{Scale the expression to a specified range c(a, b), default is NULL.}

\item{mergeMethod}{Merge method for mutiple FCS expression data. cells can be combined using one of the four different methods including \code{ceil}, \code{all}, \code{min}, \code{fixed}. The default option is 
\code{ceil}, up to a fixed number (specified by \code{fixedNum}) of cells are sampled without replacement from each fcs file and combined for analysis.
\code{all}: all cells from each fcs file are combined for analysis. 
\code{min}: The minimum number of cells among all the selected fcs files are sampled from each fcs file and combined for analysis. 
\code{fixed}: a fixed num (specified by fixedNum) of cells are sampled (with replacement when the total number of cell is less than fixedNum) from each fcs file and combined for analysis.}

\item{fixedNum}{The fixed number of cells to be extracted from each FCS file.}

\item{sampleSeed}{A sampling seed for reproducible expression matrix merging.}

\item{...}{Other arguments passed to \code{cytof_exprsExtract}}
}
\value{
A matrix containing the merged expression data, with selected markers, row names added as \code{filename_cellID}, column names added as \code{name<desc>}.
}
\description{
Apply preprocessing on each FCS file including compensation (for FCM data only) and transformation
with selected markers, then expression matrix are extracted and merged using one of the methods, 
\code{all}, \code{min}, \code{fixed} or \code{ceil}
}
\examples{
d<-system.file('extdata',package='cytofkit')
fcsFiles <- list.files(d,pattern='.fcs$',full=TRUE)
merged <- cytof_exprsMerge(fcsFiles)
}
\seealso{
\code{\link{cytof_exprsExtract}}
}
