% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_postProcess.R
\name{cytof_heatmap}
\alias{cytof_heatmap}
\title{Heatmap plot of cluster mean value results}
\usage{
cytof_heatmap(data, baseName = "Cluster", scaleMethod = "none",
  cex_row_label = NULL, cex_col_label = NULL)
}
\arguments{
\item{data}{a matrix with rownames and colnames}

\item{baseName}{The name as a prefix in the title of the heatmap.}

\item{scaleMethod}{Method indicating if the values should be centered and scaled in either the row direction or the column direction, or none. The default is 'none'.}

\item{cex_row_label}{Text size for row labels.}

\item{cex_col_label}{Text size for column labels.}
}
\value{
a heatmap object from \code{gplots}
}
\description{
Heatmap plot of cluster mean value results
}
\examples{
m1 <- c(rnorm(300, 10, 2), rnorm(400, 4, 2), rnorm(300, 7))
m2 <- c(rnorm(300, 4), rnorm(400, 16), rnorm(300, 10, 3))
m3 <- c(rnorm(300, 16), rnorm(400, 40, 3), rnorm(300, 10))
m4 <- c(rnorm(300, 7, 3), rnorm(400, 30, 2), rnorm(300, 10))
m5 <- c(rnorm(300, 27), rnorm(400, 40, 1),rnorm(300, 10))
c <- c(rep(1,300), rep(2,400), rep(3,300))
rnames <- paste(paste('sample_', c('A','B','C','D'), sep = ''), 
rep(1:250,each = 4), sep='_') 
exprs_cluster <- data.frame(cluster = c, m1 = m1, m2 = m2, m3 = m3, m4 = m4, m5 = m5)
row.names(exprs_cluster) <- sample(rnames, 1000)
cluster_mean <- aggregate(. ~ cluster, data = exprs_cluster, mean)
rownames(cluster_mean) <- paste("cluster_", cluster_mean$cluster, sep = "")
cytof_heatmap(cluster_mean[, -which(colnames(cluster_mean) == "cluster")])
}

