% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_progression.R
\name{cytof_progression}
\alias{cytof_progression}
\title{Progression estimation of cytof expression data}
\usage{
cytof_progression(data, cluster, method = "isomap",
  uniformClusterSize = 500, seed = 500)
}
\arguments{
\item{data}{Expression data matrix.}

\item{cluster}{A vector of cluster results for the data.}

\item{method}{Method for estimation of cell progression, isomap by default, tsne or pca.}

\item{uniformClusterSize}{The down sampled size of each cluster.}

\item{seed}{The seed for random down sample of the clusters.}
}
\value{
a list includes sampleData, sampleCluster and progressionData.
}
\description{
Apply isomap to estimate the relationship of cell progression
}
\examples{
data(iris)
in_data <- iris[, 1:4]
out_data <- cytof_progression(in_data, cluster = iris[,5], uniformClusterSize = 50)
}

