% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sidap_run.R
\name{cytof_tsne_densvm}
\alias{cytof_tsne_densvm}
\title{CyTOF data analysis for subpopulation detection}
\usage{
cytof_tsne_densvm(rawFCSdir = getwd(), fcsFile = NULL, resDir = getwd(),
  baseName = "sidap_analysis", para = NULL, paraFile = "./parameter.txt",
  comp = FALSE, verbose = FALSE, lgclMethod = "fixed",
  mergeMethod = "ceil", fixedNum = 10000, ifTransform = TRUE,
  transformMethod = "tsne", ifCluster = TRUE, writeResluts = TRUE, ...)
}
\arguments{
\item{rawFCSdir}{the directory that contains fcs files to be analysed.}

\item{fcsFile}{a vector containing names of fcs files to be analyzed. One or multiple fcs files are allowed.}

\item{resDir}{the directory where result files will be generated.}

\item{baseName}{a prefix that will be added to the names of result files.}

\item{para}{the vector of selected makers. This can be provided in the \code{paraFile}.}

\item{paraFile}{a text file that specifies the list of makers to be used for analysis.}

\item{comp}{Boolean tells if do compensation. This will be applied to flow cytometry data.}

\item{verbose}{Boolean.}

\item{lgclMethod}{Logicle transformation method, either \code{auto} or \code{fixed}.}

\item{mergeMethod}{when multiple fcs files are selected, cells can be combined using
one of the four different methods including \code{ceil}, \code{all}, \code{min}, \code{fixed}.
The default option is \code{ceil}, up to a fixed number (specified by \code{fixedNum}) of cells are sampled
without replacement from each fcs file and combined for analysis.
\code{all}: all cells from each fcs file are combined for analysis.
\code{min}: The minimum number of cells among all the selected fcs files are sampled from each fcs file and combined for analysis.
\code{fixed}: a fixed num (specified by fixedNum) of cells are sampled (with replacement when the total number of cell is less than
fixedNum) from each fcs file and combined for analysis.}

\item{fixedNum}{up to fixedNum of cells from each fcs file are used for analysis.}

\item{ifTransform}{a boolean to decide if dimensionality reduction will be performed. Default is TRUE.}

\item{transformMethod}{the method used for dimensionality reduction, including \code{tsne}, \code{pca}, \code{isomap}, and \code{lle}.}

\item{ifCluster}{a boolean to determine if cluster will be conducted.}

\item{writeResluts}{if save the results, and the post-processing results including scatter plot, heatmap, and statistical results.}

\item{...}{more arguments contral the logicle transformation}
}
\value{
a list containing \code{lgclMergedExprs}, \code{transData} and \code{clustersRes}. If choose 'writeResluts = TRUE', results will be saved into files under \code{resDir}
}
\description{
\code{cytof_tsne_densvm} provides a workflow for one or multiple CyToF data analysis,
including data preprocess with merging methods of multiple fcs file, logicle transformation,
dimension reduction with PCA, lle, isomap or tsne(default), and a kernal-based local maxima
clustering combined with SVM for subpopulation detection. The intermediate results can be saved
into seperate files and the cluster results can be visualized in heatmaps and scatter plots.
}
\examples{
dir <- system.file('extdata',package='sidap')
file <- list.files(dir, pattern='.fcs$', full=TRUE)
parameters <- list.files(dir, pattern='.txt$', full=TRUE)
## remove the hash symbol to run the following command
#cytof_tsne_densvm(fcsFile = file, paraFile = parameters, rawFCSdir = dir, baseName = 'test')
}
\author{
Chen Jinmiao
}
\references{
\url{http://signbioinfo.github.io/SIDAP/}
}
\seealso{
\code{\link{sidap}}, \code{\link{cytof_tsne_densvm_GUI}}
}

