% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cytof_postProcess.R
\name{cytof_write_results}
\alias{cytof_write_results}
\title{Save the cytofkit analysis results}
\usage{
cytof_write_results(analysis_results, vizMethods,
  baseName = "cytofkit_analysis", rawFCSdir = getwd(), resDir = getwd())
}
\arguments{
\item{analysis_results}{result data from output of \code{\link{DensVM_cluster}}}

\item{vizMethods}{visualization methods for clustering results, including \code{tsne}, \code{pca} and \code{isomap}.}

\item{baseName}{a prefix that will be added to the names of result files.}

\item{rawFCSdir}{the directory that contains fcs files to be analysed.}

\item{resDir}{the directory where result files will be generated.}
}
\value{
save all results in the \code{resDir}
}
\description{
Scatter dot plot and heatmap of the cluster results, and all intermediated files will be
generated and saved in the \code{resDir}
}
\examples{
dir <- system.file('extdata',package='cytofkit')
f <- list.files(dir, pattern='.fcs$', full=TRUE)
p <- list.files(dir, pattern='.txt$', full=TRUE)
#tr <- cytof_tsne_densvm(fcsFile=f,paraFile=p,baseName='t',writeResults=FALSE)
#cytof_write_results(tr,baseName = 'test',rawFCSdir=dir)
}
\seealso{
\code{\link{cytof_tsne_densvm}}, \code{\link{cytofkit}}
}

