% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_preProcess.R
\name{cytofAsinh}
\alias{cytofAsinh}
\title{Noise reduced arsinh transformation}
\usage{
cytofAsinh(value, cofactor = 5)
}
\arguments{
\item{value}{A vector of numeric values.}

\item{cofactor}{Cofactor for asinh transformation, default 5 for mass cytometry data.}
}
\value{
transformed value
}
\description{
Inverse hyperbolic sine transformation (arsinh) with a cofactor of 5, reduce noise from negative values
Adopted from Plos Comp reviewer
}

