% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cytofkit.R
\docType{package}
\name{cytofkit}
\alias{cytofkit}
\alias{cytofkit-package}
\title{cytofkit: an integrated analysis pipeline for mass cytometry data}
\description{
This package is designed to facilitate the analysis workflow of mass cytometry data with
automatic subset identification and population boundary detection. Both command line and
a GUI are provided for runing the workflow easily.
}
\details{
This package integrates merge methods of multiple FCS files, dimension reduction (PCA, t-SNE and ISOMAP)
with density-based clustering (DensVM) for rapid subset detection. Subset-clustering scatter plot
and heat map will be generated for objective comparative analysis and statistical testing. This workflow can be
easily done using the main function \code{\link{cytof_tsne_densvm}} or a GUI for the main function
\code{\link{cytof_tsne_densvm_GUI}}.

Pre-processing

Using function \code{\link{fcs_trans_merge}}, one or multiple FCS files were imported via the *read.FCS*
function in the *flowCore* package. Then logicle transformation was applied to the expression value
of selected markers of each FCS file. Auto logicle transformation and fixed logicle transformation
are provided, then mutilple FCS files are merged using method \code{all}, \code{min}, \code{fixed}
or \code{ceil}.

Dimensionality reduction

Using function \code{\link{cytof_dimReduction}}, t-Distributed Stochastic Neighbor Embedding (\code{tsne})
is suggested for dimensionality reduction although we also provide methods like \code{isomap} and \code{pca}.

Cluster analysis using DensVM

Density-based clustering aided by support Vector Machine (\code{\link{densVM_cluster}}) are used to automate
subset detection from the dimension-reducted map. By using DensVM, we are able to objectively assign every
cell to an appropriate cluster.

Post-processing

Cluster results are annotated by using scatter plot and heatmap. Scatter plot visualize the cell points
with colour indicating their assigned clusters and point shape representing their belonging samples
(\code{\link{cluster_plot}} and \code{\link{cluster_gridPlot}}). Cell events are also grouped by clusters
and samples, and mean intensity values per cluster for every marker is calculated
(\code{\link{clust_mean_heatmap}} and \code{\link{clust_percentage_heatmap}}).
Heat map visualizing the mean expression of every marker in every cluster is generated with no scaling on
the row or column direction. Hierarchical clustering was generated using Euclidean distance and complete
agglomeration method. We used the heat maps to interrogate marker expression to identify each cluster's
defining markers. All intermediate files and the plots can be saved using the function \code{\link{cytof_write_results}}.
}
\examples{
## Run on GUI
#cytof_tsne_densvm_GUI()  # remove the hash symbol to launch the GUI

## Run on command
dir <- system.file('extdata',package='cytofkit')
file <- list.files(dir, pattern='.fcs$', full=TRUE)
parameters <- list.files(dir, pattern='.txt$', full=TRUE)
## remove the hash symbol to run the following command
#cytof_tsne_densvm(fcsFile = file, paraFile = parameters, rawFCSdir = dir, baseName = 'test')

## Checking the vignettes for more details
if(interactive()) browseVignettes(package = 'cytofkit')
}
\references{
\url{http://signbioinfo.github.io/cytofkit/}
}
\seealso{
\code{\link{cytof_tsne_densvm}}, \code{\link{cytof_tsne_densvm_GUI}}
}

