% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytofkit_GUI.R
\name{cytofkit_GUI}
\alias{cytofkit_GUI}
\title{The user friendly GUI client for \code{cytofkit-package}}
\usage{
cytofkit_GUI()
}
\value{
the GUI for \code{cytofkit-package}
}
\description{
This GUI provides an easy way for CyToF data analysis using \code{cytofkit} package. 
Main parameters for running 'cytofkit' were integrated in this GUI, 
and each parameter has a help button to show the instruction. 
\code{cytofkit} analysis will be launched after submitting.
}
\examples{
#cytofkit_GUI()  # remove the hash symbol to run
}
\author{
Hao Chen
}
\references{
\url{http://signbioinfo.github.io/cytofkit/}
}
\seealso{
\code{\link{cytofkit-package}}, \code{\link{cytofkit}}
}

