% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytofkit_shinyAPP.R
\name{cytofkitShinyAPP2}
\alias{cytofkitShinyAPP2}
\title{A Shiny APP to interactively visualize the analysis results}
\usage{
cytofkitShinyAPP2(obj)
}
\arguments{
\item{objFile}{Either the RData object file or data object, if missing, RData file need to be loaded on the ShinyAPP}
}
\description{
Take the the RData object file saved by cytofkit as input, automatically load the data and allow exploration of the analysis results with interactive control
}
\examples{
if (interactive()) cytofkit::cytofkitShinyAPP2()
}
\author{
Hao Chen
}
