% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cytof_cluster.R
\name{densVM_cluster}
\alias{densVM_cluster}
\title{Density-based local maxima cluster with SVM}
\usage{
densVM_cluster(ydata, xdata)
}
\arguments{
\item{ydata}{a matrix of the dimension reduced(transformed) data}

\item{xdata}{a matrix of the expression data}
}
\value{
a list contains a matrix \code{peakdata} of the peak numbers with different kernel bandwidth, and a matrix \code{clusters} of the cluster results
}
\description{
Density-based local maxima peak finding, subpopulation assigning with the power of SVM
}
\examples{
d<-system.file('extdata',package='cytofkit')
fcsFile <- list.files(d,pattern='.fcs$',full=TRUE)
xdata <- fcs_lgcl_merge(fcsFile, mergeMethod = 'fixed', fixedNum = 100)
ydata <- cytof_dimReduction(xdata)
clusters <- densVM_cluster(ydata, xdata)
}

