% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/densityClustX.R
\name{dimReduction}
\alias{dimReduction}
\title{Dimensionality reduction for high-dimensional data}
\usage{
dimReduction(data, method = "tsne", outDim = 2)
}
\arguments{
\item{data}{Numeric matrix of data or data frame.}

\item{method}{Dimensionality reduction method, Options are: 'tsne' | 'pca'. default is 'tsne'.}

\item{outDim}{the output number of dimensionalty, default value is 2}
}
\value{
a matrix with columns of outDim, row names matching the row names of data
}
\description{
When the dimensionality of the data matrix is high, a dimensionality
reduction is suggested. A linear method PCA and a non-linear method
t-SNE are provided(suggested).
}

