% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cytof_preProcess.R
\name{fcs_lgcl}
\alias{fcs_lgcl}
\title{Logicle transformation of the FCS data}
\usage{
fcs_lgcl(fcsFile, comp = FALSE, verbose = FALSE, markers = NULL,
  lgclMethod = "fixed", w = 0.1, t = 4000, m = 4.5, a = 0)
}
\arguments{
\item{fcsFile}{The name of the FCS file}

\item{comp}{Boolean tells if do compensation}

\item{verbose}{Boolean}

\item{markers}{Selected markers for analysis, either from names or from description}

\item{lgclMethod}{Logicle transformation method, either \code{auto} or \code{fixed}}

\item{w}{Linearization width in asymptotic decades}

\item{t}{Top of the scale data value}

\item{m}{Full width of the transformed display in asymptotic decades}

\item{a}{Additional negative range to be included in the display in asymptotic decades}
}
\value{
The logicle transformend expression data matrix of selected markers
}
\description{
Read the FCS expresssion data and apply Logicle transformation
}
\examples{
d<-system.file('extdata',package='sidap')
fcsFile <- list.files(d,pattern='.fcs$',full=TRUE)
transformed <- fcs_lgcl(fcsFile)
}

