% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cytof_preProcess.R
\name{fcs_trans}
\alias{fcs_trans}
\title{Logicle transformation of the FCS data}
\usage{
fcs_trans(fcsFile, comp = FALSE, verbose = FALSE, markers = NULL,
  transformationMethod = "arcsin", scaleTo = NULL, w = 0.1, t = 4000,
  m = 4.5, a = 0, q = 0.05)
}
\arguments{
\item{fcsFile}{The name of the FCS file}

\item{comp}{Boolean tells if do compensation}

\item{verbose}{Boolean}

\item{markers}{Selected markers for analysis, either from names or from description}

\item{transformationMethod}{transformationMethod transformation method, three logicle transformation methods includes: \code{auto}, \code{sign_auto} or \code{fixed} for FCM data, and \code{arcsin} for CyTOF data}

\item{scaleTo}{scale the expression to same scale, default is NULL, should be a vector of two numbers if scale}

\item{w}{Linearization width in asymptotic decades}

\item{t}{Top of the scale data value}

\item{m}{Full width of the transformed display in asymptotic decades}

\item{a}{Additional negative range to be included in the display in asymptotic decades}

\item{q}{quantile of negative values removed for auto w estimation, default is 0.05}
}
\value{
The logicle transformend expression data matrix of selected markers
}
\description{
Read the FCS expresssion data and apply Logicle transformation
}
\examples{
d<-system.file('extdata',package='cytofkit')
fcsFile <- list.files(d,pattern='.fcs$',full=TRUE)
transformed <- fcs_trans(fcsFile)
}

