% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cytof_preProcess.R
\name{fcs_trans_merge}
\alias{fcs_trans_merge}
\title{merge the transformed expression data of FCS file(s) of selected markers}
\usage{
fcs_trans_merge(fcsFiles, comp = FALSE, verbose = FALSE, markers = NULL,
  transformationMethod = "arcsin", scaleTo = NULL, w = 0.1, t = 4000,
  m = 4.5, a = 0, q = 0.05, mergeMethod = "ceil", fixedNum = 10000)
}
\arguments{
\item{fcsFiles}{the input fcsFiles (usually more than 1 file)}

\item{comp}{Boolean tells if do compensation}

\item{verbose}{Boolean}

\item{markers}{Selected markers for analysis, either from names or from description}

\item{transformationMethod}{transformationMethod transformation method, three logicle transformation methods includes: \code{auto}, \code{sign_auto} or \code{fixed} for FCM data, and \code{arcsin} for CyTOF data}

\item{scaleTo}{scale the expression to same scale, default is NULL, should be a vector of two numbers if scale}

\item{w}{Linearization width in asymptotic decades}

\item{t}{Top of the scale data value}

\item{m}{Full width of the transformed display in asymptotic decades}

\item{a}{Additional negative range to be included in the display in asymptotic decades}

\item{q}{quantile of negative values removed for auto w estimation, default is 0.05}

\item{mergeMethod}{merge method for mutiple FCS expression data, default is all}

\item{fixedNum}{the fixed number of cells for merging multiple FCSs}
}
\value{
Merged FCS expression data matrix of selected markers with logicle transformation
}
\description{
Apply logicle transformation of selected markers of each FCS file, auto logicle
transformation and fixed logicle transformation are provided, then mutilple
FCS files are merged using method \code{all}, \code{min}, \code{fixed} or \code{ceil}
}
\examples{
d<-system.file('extdata',package='cytofkit')
fcsFile <- list.files(d,pattern='.fcs$',full=TRUE)
merged <- fcs_trans_merge(fcsFile)
}

