% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytofkit_GUI.R
\name{launchShinyAPP_GUI}
\alias{launchShinyAPP_GUI}
\title{GUI for launching shiny APP}
\usage{
launchShinyAPP_GUI(message = "cytofkit", dir = getwd())
}
\arguments{
\item{message}{A message to determine if open the shiny APP}

\item{dir}{Result direcroty.}
}
\description{
A shiny APP for interactive exploration of the analysis results
}
\examples{
# launchShinyAPP_GUI()
}

