% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/densityClustX.R
\name{minDistToHigher}
\alias{minDistToHigher}
\title{Calculate distance to closest observation of higher density}
\usage{
minDistToHigher(data, rho, distMethod, ifParallel = parallel)
}
\arguments{
\item{data}{Numeric matrix of data or data frame.}

\item{rho}{A vector of local density values as outputted by \code{localDensity}}

\item{distMethod}{the method used for distance calculation}

\item{ifParallel}{A boolean decides if run parallelly}
}
\value{
A list of distances to closest observation of higher density and the ID
}
\description{
This function finds, for each observation, the minimum distance to an
observation of higher local density. With a rowise implementation of
the pairwise distance calculation, makes the local density estimation
faster and memory efficient. A big benifit is the aviliability for big
data. Parallel computing is supported for fast calculation.
}

