% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cytof_postProcess.R
\name{peaksGamma_plot}
\alias{peaksGamma_plot}
\title{Plot varaition of peak nums with increasing gamma}
\usage{
peaksGamma_plot(peakdata)
}
\arguments{
\item{peakdata}{a matrix of \code{peakdata} returned from \code{DensVM_cluster}}
}
\value{
a line graph of peak nums vs. increasing gamma
}
\description{
Plot varaition of peak nums with increasing gamma
}
\examples{
x <- seq(0, 1, length.out = 20)
y <- c(20:6, 6, 6, 5:3)
peakdata <- data.frame(sig_range = x, numpeaks = y)
peaksGamma_plot(peakdata)
}

