% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rphenograph.R
\name{Rphenograph}
\alias{Rphenograph}
\title{RphenoGraph clustering}
\usage{
Rphenograph(data, k = 30)
}
\arguments{
\item{data}{Input data matrix.}

\item{k}{Number of nearest neighbours, default is 30.}
}
\value{
a communities object, the operations of this class contains:
\item{print}{returns the communities object itself, invisibly.}
\item{length}{returns an integer scalar.}
\item{sizes}{returns a numeric vector.}
\item{membership}{returns a numeric vector, one number for each vertex in the graph that was the input of the community detection.}
\item{modularity}{returns a numeric scalar.}
\item{algorithm}{returns a character scalar.}
\item{crossing}{returns a logical vector.}
\item{is_hierarchical}{returns a logical scalar.}
\item{merges}{returns a two-column numeric matrix.}
\item{cut_at}{returns a numeric vector, the membership vector of the vertices.}
\item{as.dendrogram}{returns a dendrogram object.}
\item{show_trace}{returns a character vector.}
\item{code_len}{returns a numeric scalar for communities found with the InfoMAP method and NULL for other methods.}
\item{plot}{for communities objects returns NULL, invisibly.}
}
\description{
R implementation of the phenograph algorithm
}
\details{
A simple R implementation of the phenograph [PhenoGraph](http://www.cell.com/cell/abstract/S0092-8674(15)00637-6) algorithm, 
which is a clustering method designed for high-dimensional single-cell data analysis. It works by creating a graph ("network") representing 
phenotypic similarities between cells by calculating the Jaccard coefficient between nearest-neighbor sets, and then identifying communities 
using the well known [Louvain method](https://sites.google.com/site/findcommunities/) in this graph.
}
\examples{
iris_unique <- unique(iris) # Remove duplicates
data <- as.matrix(iris_unique[,1:4])
Rphenograph_out <- Rphenograph(data, k = 45)
}
\references{
Jacob H. Levine and et.al. Data-Driven Phenotypic Dissection of AML Reveals Progenitor-like Cells that Correlate with Prognosis. Cell, 2015.
}
\author{
Chen Hao
}
