% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_preProcess.R
\name{scaleData}
\alias{scaleData}
\title{rescale the data}
\usage{
scaleData(x, range = c(0, 4.5))
}
\arguments{
\item{x}{data.}

\item{range}{The range of the data.}
}
\value{
scaled data
}
\description{
rescale the data
}

