% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_postProcess.R
\name{spectral1}
\alias{spectral1}
\title{First Function for spectral color palette}
\usage{
spectral1(n)
}
\arguments{
\item{n}{Number of colors.}
}
\value{
Hex colour values
}
\description{
First Function for spectral color palette
}
\examples{
spectral1(2)
}
