% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cytof_preProcess.R
\name{applyComp}
\alias{applyComp}
\title{apply compensation on the FCS expression data}
\usage{
applyComp(fcs, keyword)
}
\arguments{
\item{fcs}{FCS file.}

\item{keyword}{Keywords.}
}
\description{
apply compensation on the FCS expression data
}

