% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cytof_preProcess.R
\name{auto_lgcl}
\alias{auto_lgcl}
\title{a modified version of "estimateLogicle" from flowCore}
\usage{
auto_lgcl(x, channels, m = 4.5, q = 0.05)
}
\description{
a modified version of "estimateLogicle" from flowCore
}

