% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterX.R
\name{ClusterX}
\alias{ClusterX}
\title{Fast clustering by automatic search and find of density peaks}
\usage{
ClusterX(data, dimReduction = NULL, outDim = 2, dc, gaussian = TRUE,
  alpha = 0.001, detectHalos = FALSE, SVMhalos = FALSE,
  parallel = FALSE, nCore = 4)
}
\arguments{
\item{data}{A data matrix for clustering.}

\item{dimReduction}{Dimenionality reduction method.}

\item{outDim}{Number of dimensions used for clustering.}

\item{dc}{Distance cutoff value.}

\item{gaussian}{If \code{TRUE}, apply gaussian to estimate the density.}

\item{alpha}{Significance level for peak detection.}

\item{detectHalos}{If \verb{TRUE}, detect the halos.}

\item{SVMhalos}{If \code{TRUE}, Run SVM on cores to assign halos.}

\item{parallel}{If \code{TRUE}, run the algorithm in parallel.}

\item{nCore}{Number of cores umployed for parallel compution.}
}
\value{
A list
}
\description{
This package implements the clustering algorithm described by Alex Rodriguez
and Alessandro Laio (2014) with improvements of automatic peak detection and
parallel implementation
}
\details{
ClusterX works on low dimensional data analysis (Dimensionality less than 5).
If input data is high dimensional, t-SNE is conducted to reduce the dimensionality.
}
\examples{
iris_unique <- unique(iris) # Remove duplicates
data <- as.matrix(iris_unique[,1:4])
ClusterXRes <- ClusterX(data)
}
\author{
Chen Hao
}
