% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/densityClustX.R
\name{ClusterX}
\alias{ClusterX}
\title{Fast clustering by automaticly search and find of density peaks}
\usage{
ClusterX(data, dimReduction = NULL, outDim = 2, distMethod = "euclidean",
  dc, gaussian = TRUE, alpha = 0.001, detectHalos = FALSE,
  parallel = FALSE, nCore = 4)
}
\arguments{
\item{data}{a data matrix for clustering}

\item{dimReduction}{dimenionality reduciton method}

\item{outDim}{number of dimensions will be used}

\item{distMethod}{distance calculaiton method, by default is euclidean}

\item{dc}{distance cutoff value}

\item{gaussian}{if apply gaussian to esitmate the density}

\item{alpha}{signance level for peak detection}

\item{detectHalos}{if detect the halos}

\item{parallel}{if run the algorithm in parallel}

\item{number}{of cores umployed for parallel compution}
}
\value{
a object of \code{densityClusterX} class
}
\description{
This package implement the clustering algorithm described by Alex Rodriguez
and Alessandro Laio (2014) with improvements of automatic peak detection and parallel implementation
}
\examples{
iris_unique <- unique(iris) # Remove duplicates
data <- as.matrix(iris_unique[,1:4])
densityClustRes <- densityClustX(data)
}

