% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_cluster.R
\name{cytof_cluster}
\alias{cytof_cluster}
\title{Subset detection by clustering}
\usage{
cytof_cluster(ydata = NULL, xdata = NULL, method = "densVM")
}
\arguments{
\item{ydata}{a matrix of the dimension reduced(transformed) data}

\item{xdata}{a matrix of the expression data}

\item{method}{cluster method including \code{densVM}, \code{densityClustX} and \code{Rphenograph}.}
}
\value{
a vector of the clusters assigned for each row of the ydata
}
\description{
Apply clustering algorithms to detect cell subsets. \code{densVM} and \code{ClusterX} clustering is 
based on the transformend ydata; \code{Rphenograph} is directly applied on the high dimemnional xdata. And 
\code{densVM} need the xdata to train the VM model.
}
\examples{
d<-system.file('extdata', package='cytofkit')
fcsFile <- list.files(d, pattern='.fcs$', full=TRUE)
xdata <- cytof_exprsMerge(fcsFile, mergeMethod = 'fixed', fixedNum = 100)
ydata <- cytof_dimReduction(xdata, method = "tsne")
clusters <- cytof_cluster(ydata, xdata, method = "densVM")
}

