% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_dimensionReduction.R
\name{cytof_dimReduction}
\alias{cytof_dimReduction}
\title{Dimension reduction for high dimensional data}
\usage{
cytof_dimReduction(data, method = c("tsne", "pca", "isomap", "NULL"),
  out_dim = 2, isomap_distMethod = "euclidean", isomap_k = 5,
  isomap_ndim = NULL, isomapFragmentOK = TRUE, ...)
}
\arguments{
\item{data}{Input expression data matrix.}

\item{method}{Method chosed for dimensition reduction, must be one of \code{isomap}, \code{pca} or \code{tsne}.}

\item{out_dim}{The dimensionality of the output.}

\item{isomap_distMethod}{Method for distance calcualtion for \code{isomap}.}

\item{isomap_k}{Number of shortest dissimilarities retained for a point, parameter for \code{isomap} method.}

\item{isomap_ndim}{Number of axes in metric scaling, parameter for \code{isomap} method.}

\item{isomapFragmentOK}{What to do if dissimilarity matrix is fragmented, parameter for \code{isomap} method.}

\item{...}{Other parameters passed to Rtsne}
}
\value{
a matrix of the dimension reducted data, with colnames method_ID, and rownames same as the input data.
}
\description{
Apply dimension reduction on the cytof expression data, 
with method \code{pca}, \code{tsne} or \code{isomap}.
}
\examples{
data(iris)
in_data <- iris[, 1:4]
out_data <- cytof_dimReduction(in_data, method = "tsne")
}

