% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_preProcess.R
\name{cytof_exprsMerge}
\alias{cytof_exprsMerge}
\title{Merge the transformed expression data of FCS file(s) of selected markers}
\usage{
cytof_exprsMerge(fcsFiles, comp = FALSE, verbose = FALSE, markers = NULL,
  transformMethod = "auto_lgcl", scaleTo = NULL, mergeMethod = "ceil",
  fixedNum = 10000, w = 0.1, t = 4000, m = 4.5, a = 0, q = 0.05)
}
\arguments{
\item{fcsFiles}{the input fcsFiles (usually more than 1 file)}

\item{comp}{Boolean value tells if do compensation}

\item{verbose}{Boolean value detecides if print the massage details}

\item{markers}{Selected markers for analysis, either from names or from description}

\item{transformMethod}{transformation method, \code{auto_lgcl}, \code{fixed_lgcl}, \code{arcsin} or \code{biexp}}

\item{scaleTo}{scale the expression to same scale, default is NULL, should be a vector of two numbers if scale}

\item{mergeMethod}{merge method for mutiple FCS expression data. cells can be combined using 
one of the four different methods including \code{ceil}, \code{all}, \code{min}, \code{fixed}. 
The default option is \code{ceil}, up to a fixed number (specified by \code{fixedNum}) of cells are sampled 
without replacement from each fcs file and combined for analysis.
\code{all}: all cells from each fcs file are combined for analysis. 
\code{min}: The minimum number of cells among all the selected fcs files are sampled from each fcs file and combined for analysis. 
\code{fixed}: a fixed num (specified by fixedNum) of cells are sampled (with replacement when the total number of cell is less than 
fixedNum) from each fcs file and combined for analysis.}

\item{fixedNum}{the fixed number of cells for merging multiple FCSs}

\item{w}{Linearization width in asymptotic decades}

\item{t}{Top of the scale data value}

\item{m}{Full width of the transformed display in asymptotic decades}

\item{a}{Additional negative range to be included in the display in asymptotic decades}

\item{q}{quantile of negative values removed for auto w estimation, default is 0.05}
}
\value{
Merged FCS expression data matrix of selected markers after transformation
}
\description{
Apply transformation of selected markers of each FCS file, arcsin, biexponential, auto logicle
transformation and fixed logicle transformation are provided, then mutilple 
FCS files are merged using method \code{all}, \code{min}, \code{fixed} or \code{ceil}
}
\examples{
d<-system.file('extdata',package='cytofkit')
fcsFiles <- list.files(d,pattern='.fcs$',full=TRUE)
merged <- cytof_exprsMerge(fcsFiles)
}

