% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_progression.R
\name{cytof_progression}
\alias{cytof_progression}
\title{Progression estimation of cytof expression data}
\usage{
cytof_progression(data, cluster, method = c("diffusionmap", "isomap", "NULL"),
  distMethod = "euclidean", out_dim = 2, clusterSampleMethod = c("ceil",
  "all", "fixed", "min"), clusterSampleSize = 500, sampleSeed = 123)
}
\arguments{
\item{data}{Expression data matrix.}

\item{cluster}{A vector of cluster results for the data.}

\item{method}{Method for estimation of cell progression, isomap or diffusionmap.}

\item{distMethod}{Method for distance calcualtion, default is "euclidean", other choices like "manhattan", "cosine", "rankcor".}

\item{out_dim}{Number of transformed dimenions choosed for output.}

\item{clusterSampleMethod}{Cluster sampling method including \code{ceil}, \code{all}, \code{min}, \code{fixed}. The default option is 
\code{ceil}, up to a fixed number (specified by \code{fixedNum}) of cells are sampled without replacement from each cluster and combined for analysis.
\code{all}: all cells from each cluster are combined for analysis. 
\code{min}: The minimum number of cells among all clusters are sampled from cluster and combined for analysis. 
\code{fixed}: a fixed num (specified by fixedNum) of cells are sampled (with replacement when the total number of cell is less than fixedNum) from each cluster and combined for analysis.}

\item{clusterSampleSize}{The number of cells to be sampled from each cluster.}

\item{sampleSeed}{The seed for random down sample of the clusters.}
}
\value{
a list includes sampleData, sampleCluster and progressionData.
}
\description{
Infer the progression based on the relationship of cell subsets estimated 
using ISOMAP or Diffusion map.
}
\examples{
d<-system.file('extdata', package='cytofkit')
fcsFile <- list.files(d, pattern='.fcs$', full=TRUE)
parameters <- list.files(d, pattern='.txt$', full=TRUE)
markers <- as.character(read.table(parameters, sep = "\\t", header = TRUE)[, 1])
xdata <- cytof_exprsMerge(fcsFile, markers = markers, mergeMethod = 'fixed', fixedNum = 2000)
clusters <- cytof_cluster(xdata = xdata, method = "Rphenograph")
prog <- cytof_progression(data = xdata, cluster = clusters, clusterSampleSize = 100)
d <- as.data.frame(cbind(prog$progressionData, cluster = factor(prog$sampleCluster)))
cytof_clusterPlot(data =d, xlab = "diffusionmap_1", ylab="diffusionmap_2", cluster = "cluster", sampleLabel = FALSE)
}
