% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cytof_progression.R
\name{cytof_progression}
\alias{cytof_progression}
\title{Progression estimation of cytof expression data}
\usage{
cytof_progression(data, cluster, method = "isomap",
  uniformClusterSize = 500, seed = 500)
}
\arguments{
\item{data}{Expression data matrix}

\item{method}{A vector of cluster results for the \code{data}}

\item{uniformClusterSize}{the down sampled size of each cluster}

\item{seed}{the seed for random down sample of the clusters}

\item{method}{Method for estimation of cell progression, \code{isomap} (by default), \code(tsne) or \code(pca)}
}
\value{
a list of the progression data including \code{sampleData}, \code{sampleCluster} and \code{progressionData}.
}
\description{
Apply isomap to estimate the order of cell progression
}
\examples{
data(iris)
in_data <- iris[, 1:4]
out_data <- cytof_progression(in_data, cluster = iris[,5], uniformClusterSize = 50)
}
\author{
Chen Jinmiao
}

