% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cytofkit_GUI.R
\name{cytof_tsne_densvm_GUI}
\alias{cytof_tsne_densvm_GUI}
\title{The user friendly GUI for function \code{cytof_tsne_densvm}}
\usage{
cytof_tsne_densvm_GUI()
}
\value{
the GUI for the main function \code{cytof_tsne_densvm}
}
\description{
This GUI provides an easy way for CyToF data analysis using \code{cytofkit} package. All parameters
for running 'cytof_tsne_densvm' were integrated in this GUI, each parameter has help button
in the GUI to help user get details of the information of each parameter, and launch the
\code{cytof_tsne_densvm} analysis after submitting.
}
\examples{
#cytof_tsne_densvm_GUI()  # remove the comment hash to run
}
\author{
Chen Hao
}
\references{
\url{http://signbioinfo.github.io/cytofkit/}
}
\seealso{
\code{\link{cytof_tsne_densvm}}, \code{\link{cytofkit}}
}

