% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_postProcess.R
\name{cytof_writeResults}
\alias{cytof_writeResults}
\title{Save the cytofkit analysis results}
\usage{
cytof_writeResults(analysis_results, projectName, resultDir, saveToFCS = TRUE,
  rawFCSdir = getwd())
}
\arguments{
\item{analysis_results}{result data from output of \code{\link{cytofkit}}}

\item{projectName}{a prefix that will be added to the names of result files.}

\item{resultDir}{the directory where result files will be generated.}

\item{saveToFCS}{save the results back to the FCS files, new FCS files will be generated.}

\item{rawFCSdir}{the directory that contains fcs files to be analysed.}
}
\value{
save all results in the \code{resultDir}
}
\description{
Scatter dot plot and heatmap of the cluster results, and all intermediate files will be 
generated and saved in the \code{resultDir}
}
\examples{
d <- system.file('extdata',package='cytofkit')
f <- list.files(d, pattern='.fcs$', full=TRUE)
p <- list.files(d, pattern='.txt$', full=TRUE)
#tr <- cytofkit(fcsFile=f,markers=p,projectName='t',saveResults=FALSE)
#cytof_write_results(tr,projectName = 'test',resultDir=d,rawFCSdir =d)
}
\seealso{
\code{\link{cytofkit}}
}

