% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cytof_postProcess.R
\name{cytof_writeResults}
\alias{cytof_writeResults}
\title{Save the cytofkit analysis results}
\usage{
cytof_writeResults(analysis_results, projectName = "cytofkit",
  resultDir = getwd(), saveToFCS = TRUE, rawFCSdir = getwd())
}
\arguments{
\item{analysis_results}{result data from output of \code{\link{cytofkit}}}

\item{projectName}{a prefix that will be added to the names of result files.}

\item{resultDir}{the directory where result files will be generated.}

\item{saveToFCS}{save the results back to the FCS files, new FCS files will be generated.}

\item{rawFCSdir}{the directory that contains fcs files to be analysed.}
}
\value{
save all results in the \code{resultDir}
}
\description{
Scatter dot plot and heatmap of the cluster results, and all intermediate files will be
generated and saved in the \code{resultDir}
}
\examples{
dir <- system.file('extdata',package='cytofkit')
f <- list.files(dir, pattern='.fcs$', full=TRUE)
p <- list.files(dir, pattern='.txt$', full=TRUE)
#tr <- cytof_tsne_densvm(fcsFile=f,paraFile=p,baseName='t',writeResults=FALSE)
#cytof_write_results(tr,baseName = 'test',rawFCSdir=dir)
}
\seealso{
\code{\link{cytof_tsne_densvm}}, \code{\link{cytofkit}}
}

