% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytofkit.R
\docType{package}
\name{cytofkit-package}
\alias{cytofkit-package}
\title{cytofkit: an integrated mass cytometry data analysis pipeline}
\description{
This package is designed to facilitate the analysis workflow of mass cytometry data with 
automatic subset identification and mapping of cellular progression. Both command line and 
a GUI client are provided for executing the workflow easily.
}
\details{
This package integrates merging methods of multiple FCS files, dimension reduction methods (PCA, t-SNE and ISOMAP) 
and clustering methods (DensVM, ClusterX, and Rphenograph) for rapid subset detection. Analysis reuslts can be visualized 
and explored interactively using a specially designed shiny web APP, see \code{\link{cytofkitShinyAPP}}. Moreover, the method isomap is provided to map the cellular progression. 
This workflow can be easily executed with the main function \code{\link{cytofkit}} or through the GUI client \code{\link{cytofkit_GUI}}.

Pre-processing

Using function \code{\link{cytof_exprsMerge}}, one or multiple FCS files will be loaded via the *read.FCS* 
function in the *flowCore* package. Then transformation was applied to the expression value 
of selected markers of each FCS file. Transformation methods include \code{autoLgcl}, \code{cytofAsinh}, 
\code{logicle} and \code{arcsinh}, where \code{cytofAsinh} is the default.Then mutilple FCS files are 
merged using one of the merging methods \code{all}, \code{min}, \code{fixed} or \code{ceil}.

Dimensionality reduction

Using function \code{\link{cytof_dimReduction}}, t-Distributed Stochastic Neighbor Embedding (\code{tsne}) 
is suggested for dimensionality reduction although we also provide methods like \code{isomap} and \code{pca}.

Cluster

Using function \code{\link{cytof_cluster}}, three cluster method are provided, \code{DensVM}, \code{ClusterX} 
\code{Rphenograph} and \code{FlowSOM}. \code{DensVM}, \code{densityClustX} are performend on the dimension reduced data, while \code{Rphenograph}
works directed on the high dimensional expression data. Method \code{FlowSOM} is integrated from FlowSOM package 
(https://bioconductor.org/packages/release/bioc/html/FlowSOM.html).

Post-processing

- Using function \code{\link{cytof_clusterPlot}} to visualize the cluster results in a catter plot, in which dots represent cells, colours 
indicate their assigned clusters and point shapes represent their belonging samples.

- Using function \code{\link{cytof_heatmap}} to generate heat map to visualize the mean expression of every marker in every cluster. 
 This heat maps is useful to interrogate marker expression to identify each cluster's defining markers.

- Using function \code{\link{cytof_progressionPlot}} to visualize the expression patter of selected markers against the estimated 
 cellular progression order.
 
 - Using function \code{\link{cytof_addToFCS}} to add any dimension reduced data, cluster results, progression data into the original FCS files,
 new FCS files will be saved for easy checking with other softwares like FlowJo.

All the above post processing can be automatically implemented and saved using one function \code{\link{cytof_writeResults}}.
}
\examples{

## Run on GUI
#cytofkit_GUI()  # remove the hash symbol to launch the GUI

## Run on command
dir <- system.file('extdata',package='cytofkit')
file <- list.files(dir, pattern='.fcs$', full=TRUE)
parameters <- list.files(dir, pattern='.txt$', full=TRUE)
## remove the hash symbol to run the following command
#cytofkit(fcsFile = file, markers = parameters)   

## Checking the vignettes for more details 
if(interactive()) browseVignettes(package = 'cytofkit')

}
\references{
\url{http://signbioinfo.github.io/cytofkit/}
}
\seealso{
\code{\link{cytofkit}}, \code{\link{cytofkit_GUI}}
}
\author{
Hao Chen, Jinmiao Chen
}
