% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cytofkit.R
\name{cytofkit}
\alias{cytofkit}
\title{cytofkit: an integrated analysis pipeline for mass cytometry data}
\usage{
cytofkit(fcsFiles = getwd(), markers = NULL, mergeMethod = "ceil",
  fixedNum = 10000, comp = FALSE, verbose = FALSE,
  transformMethod = "auto_lgcl", scaleTo = NULL, q = 0.05,
  dimReductionMethod = "tsne", clusterMethods = "densVM",
  visualizationMethods = "tsne", progressionMethod = NULL,
  uniformClusterSize = 500, projectName = "cytofkit", resultDir = getwd(),
  saveResults = TRUE, saveObject = TRUE, saveToFCS = TRUE, ...)
}
\arguments{
\item{fcsFiles}{it can be either the name of the path where stores your FCS files or a vector of FCS file names.}

\item{markers}{it can be either a text file that specifies the makers to be used for analysis or a vector of the marker names.}

\item{mergeMethod}{when multiple fcs files are selected, cells can be combined using
one of the four different methods including \code{ceil}, \code{all}, \code{min}, \code{fixed}.
The default option is \code{ceil}, up to a fixed number (specified by \code{fixedNum}) of cells are sampled
without replacement from each fcs file and combined for analysis.
\code{all}: all cells from each fcs file are combined for analysis.
\code{min}: The minimum number of cells among all the selected fcs files are sampled from each fcs file and combined for analysis.
\code{fixed}: a fixed num (specified by fixedNum) of cells are sampled (with replacement when the total number of cell is less than
fixedNum) from each fcs file and combined for analysis.}

\item{fixedNum}{up to fixedNum of cells from each fcs file are used for analysis.}

\item{comp}{Boolean value to decide if do compensation. This will be applied to flow cytometry data.}

\item{verbose}{Boolean value, if detailed massaged will be printed.}

\item{transformMethod}{dat transformation method, either \code{auto_lgcl}, \code{fixed_lgcl}, \code{arcsin} or \code{biexp}.}

\item{scaleTo}{scale the expression values to the same scale, default is NULL, should be a vector of two numbers if scale.}

\item{q}{quantile of negative values removed for auto w estimation in logicle transformation, default is 0.05.}

\item{dimReductionMethod}{the method used for dimensionality reduction, including \code{tsne}, \code{pca} and \code{isomap}.}

\item{clusterMethods}{the clustering method(s) used for subpopulation detection, including \code{densVM}, \code{ClusterX} and \code{Rphenograph}. Multiple selection are accepted.}

\item{visualizationMethods}{the method(s) used for visualize the cluster data, including \code{tsne}, \code{pca} and \code{isomap}. Multiple selection are accepted.}

\item{progressionMethod}{use the first ordination score of \code{isomap} to estimated the preogression order of cells, choose \code{NULL} to ignore.}

\item{uniformClusterSize}{the uniform size of each cluster.}

\item{projectName}{a prefix that will be added to the names of result files.}

\item{resultDir}{the directory where result files will be generated.}

\item{saveResults}{if save the results, and the post-processing results including scatter plot, heatmap, and statistical results.}

\item{saveObject}{save the resutls into RData objects for loading back to R for further analysis}

\item{saveToFCS}{save the results back to the FCS files, new FCS files will be generated.}

\item{...}{more arguments contral the logicle transformation}
}
\value{
a list containing \code{expressionData}, \code{dimReductionMethod}, \code{visualizationMethods}, \code{dimReducedRes}, \code{clusterRes} and \code{progressionRes}. If choose 'saveResults = TRUE', results will be saved into files under \code{resultDir}
}
\description{
\code{cytofkit} provides a workflow for one or multiple cytof data analysis,
including data preprocess with merging methods of multiple fcs file, expression data transformation,
dimension reduction with PCA, isomap or tsne(default), clustering methods(densVM, ClusterX, Rphenograph)
for subpopulation detection, and estimation of cellular progression with isomap. The analysis results
can be visualized with scatter plot, heatmap plot or progression plot. Moreover theses results can be saved back to
FCS files. By default the results will be automatically saved for further annoration.
}
\examples{
dir <- system.file('extdata',package='cytofkit')
file <- list.files(dir, pattern='.fcs$', full=TRUE)
parameters <- list.files(dir, pattern='.txt$', full=TRUE)
## remove the hash symbol to run the following command
#cytofkit(fcsFile = file, markers = parameters, projectName = 'test')
}
\author{
Chen Jinmiao, Chen Hao
}
\references{
\url{http://signbioinfo.github.io/cytofkit/}
}
\seealso{
\code{\link{cytofkit}}, \code{\link{cytofkit_GUI}}
}

