% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytofkit_GUI.R
\name{cytofkit_GUI}
\alias{cytofkit_GUI}
\title{Function for launching the user friendly GUI client for \code{cytofkit-package}}
\usage{
cytofkit_GUI()
}
\value{
the GUI for \code{cytofkit-package}
}
\description{
This GUI provides an easy way to apply \code{cytofkit} package. 
Main parameters for running 'cytofkit' main function were integrated in this GUI, 
and each parameter has a help button to show the instructions. 
The \code{cytofkit} analysis will be automatically started after submitting.
}
\examples{
#cytofkit_GUI()  # remove the hash symbol to run
}
\references{
\url{http://signbioinfo.github.io/cytofkit/}
}
\seealso{
\code{\link{cytofkit-package}}, \code{\link{cytofkit}}
}
\author{
Hao Chen
}
