% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cytofkit.R
\name{cytofkitShinyAPP}
\alias{cytofkitShinyAPP}
\title{A Shiny app to interactively visualize the analysis results}
\usage{
cytofkitShinyAPP()
}
\description{
A Shiny app to interactively visualize the analysis results
}
\examples{
if (interactive()) cytofkit::cytofkitShinyAPP()
}

