% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytofkit.R
\name{cytofkitShinyAPP}
\alias{cytofkitShinyAPP}
\title{A Shiny app to interactively visualize the analysis results}
\usage{
cytofkitShinyAPP()
}
\description{
Load the RData object saved by cytofkit, explore the analysis results with interactive control
}
\examples{
if (interactive()) cytofkit::cytofkitShinyAPP()
}
\author{
Hao Chen
}

