% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytofkit_GUI.R
\name{launchShinyAPP_GUI}
\alias{launchShinyAPP_GUI}
\title{GUI for launching shiny APP}
\usage{
launchShinyAPP_GUI(message = "cytofkit", dir = getwd(), obj = NULL)
}
\arguments{
\item{message}{Message when asking if user wants to open the shiny APP}

\item{dir}{Result directory.}

\item{obj}{The RData piped from cytofkit function to the Shiny App}
}
\value{
Window asking user if they wish to open shinyApp directly
}
\description{
A shiny APP for interactive exploration of analysis results
}
\examples{
# launchShinyAPP_GUI()
}
