% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytof_postProcess.R
\name{spectral2}
\alias{spectral2}
\title{Second Function for spectral color palette}
\usage{
spectral2(n)
}
\arguments{
\item{n}{Number of colors.}
}
\value{
Hex colour values
}
\description{
Second Function for spectral color palette
}
\examples{
spectral2(2)
}
